create SlidingWindowGroupByExecutorFactory
